Populate source_map
Generated an example source to link to the observations in the loader.  Used the example source_id to populate the source_map table.  Future work would be to create a source loading step that loaded sources from a data file.

In the process of adding this feature we discovered that the load process was loading objects with no associated observations for craters that had a unicodedecodeerror in the crater data.  We corrected this by adding a transaction rollback in the except block of the load step to clear the pending object creation transaction before moving on to the next crater.  The rollback prevent the object from being inserted in the next commit.